struttura dati modificata

+++ struttura_dati.py
@@ -57,8 +57,8 @@ PERSONAGGI = [
                 "y_barca": (HEIGHT // 2) - (HEIGHT // 16),
             },
             "main": {
-                "x_attuale": random.randint(400*MOD, WIDTH - 420*MOD),
-                "y_attuale": random.randint(430*MOD, 470*MOD),
+                "x_attuale": random.randint(int(400*MOD), int(WIDTH - 420*MOD)),
+                "y_attuale": random.randint(int(430*MOD), int(470*MOD)),
             },
         },
         "sprites": {
@@ -87,8 +87,8 @@ PERSONAGGI = [
                 "y_barca": (HEIGHT // 2) - (HEIGHT // 16) - (20 * MOD),
             },
             "main": {
-                "x_attuale": random.randint(400*MOD, WIDTH - 420*MOD),
-                "y_attuale": random.randint(430*MOD, 470*MOD),
+                "x_attuale": random.randint(int(400*MOD), int(WIDTH - 420*MOD)),
+                "y_attuale": random.randint(int(430*MOD), int(470*MOD)),
             },
         },
         "sprites": {
@@ -117,8 +117,8 @@ PERSONAGGI = [
                 "y_barca": (HEIGHT // 2) - (HEIGHT // 16) - (20 * MOD),
             },
             "main": {
-                "x_attuale": random.randint(400*MOD, WIDTH - 420*MOD),
-                "y_attuale": random.randint(430*MOD, 470*MOD),
+                "x_attuale": random.randint(int(400*MOD), int(WIDTH - 420*MOD)),
+                "y_attuale": random.randint(int(430*MOD), int(470*MOD)),
             },
         },
         "sprites": {
@@ -147,8 +147,8 @@ PERSONAGGI = [
                 "y_barca": (HEIGHT // 2) - (HEIGHT // 16) - (50 * MOD),
             },
             "main": {
-                "x_attuale": random.randint(400*MOD, WIDTH - 420*MOD),
-                "y_attuale": random.randint(430*MOD, 470*MOD),
+                "x_attuale": random.randint(int(400*MOD), int(WIDTH - 420*MOD)),
+                "y_attuale": random.randint(int(430*MOD), int(470*MOD)),
             },
         },
         "sprites": {
@@ -177,8 +177,8 @@ PERSONAGGI = [
                 "y_barca": (HEIGHT // 2) - (HEIGHT // 16) - (40 * MOD),
             },
             "main": {
-                "x_attuale": random.randint(400*MOD, WIDTH - 420*MOD),
-                "y_attuale": random.randint(430*MOD, 470*MOD),
+                "x_attuale": random.randint(int(400*MOD), int(WIDTH - 420*MOD)),
+                "y_attuale": random.randint(int(430*MOD), int(470*MOD)),
             },
         },
         "sprites": {
@@ -207,8 +207,8 @@ PERSONAGGI = [
                 "y_barca": (HEIGHT // 2) - (HEIGHT // 16),
             },
             "main": {
-                "x_attuale": random.randint(400*MOD, WIDTH - 420*MOD),
-                "y_attuale": random.randint(430*MOD, 470*MOD),
+                "x_attuale": random.randint(int(400*MOD), int(WIDTH - 420*MOD)),
+                "y_attuale": random.randint(int(430*MOD), int(470*MOD)),
             },
         },
         "sprites": {
@@ -237,8 +237,8 @@ PERSONAGGI = [
                 "y_barca": (HEIGHT // 2) - (HEIGHT // 16) - (10 * MOD),
             },
             "main": {
-                "x_attuale": random.randint(400*MOD, WIDTH - 420*MOD),
-                "y_attuale": random.randint(430*MOD, 470*MOD),
+                "x_attuale": random.randint(int(400*MOD), int(WIDTH - 420*MOD)),
+                "y_attuale": random.randint(int(430*MOD), int(470*MOD)),
                 
             },
         },
@@ -268,8 +268,8 @@ PERSONAGGI = [
                 "y_barca": (HEIGHT // 2) - (HEIGHT // 16) - (79 * MOD),
             },
             "main": {
-                "x_attuale": random.randint(400*MOD, WIDTH - 420*MOD),
-                "y_attuale": random.randint(430*MOD, 470*MOD),
+                "x_attuale": random.randint(int(400*MOD), int(WIDTH - 420*MOD)),
+                "y_attuale": random.randint(int(430*MOD), int(470*MOD)),
                 
             },
         },
